requirements.txt
pdfplumber
sentence-transformers
scikit-learn
google-generativeai
python-dotenv
gunicorn
Django
tailwind